Moved test/util to the testutil folder
