+++ network/overlay_test.go
@@ -28,7 +28,7 @@ import (
 	"github.com/iotexproject/iotex-core/config"
 	"github.com/iotexproject/iotex-core/network/node"
 	"github.com/iotexproject/iotex-core/proto"
-	"github.com/iotexproject/iotex-core/test/util"
+	"github.com/iotexproject/iotex-core/testutil"
 )
 
 func LoadTestConfig(addr string, allowMultiConnsPerIP bool) *config.Network {
